+++ model/schema/imsglobal/classification/LomClassificationSchemaTrait.php
@@ -14,48 +14,26 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2017 Open Assessment Technologies SA
+ * Copyright (c) 2017 (original work) Open Assessment Technologies SA
  *
  */
 
 namespace oat\taoLom\model\schema\imsglobal\classification;
 
-use oat\taoLom\model\schema\LomMetadataAbstract;
 
-class LomClassificationEntryMetadata extends LomMetadataAbstract
+trait LomClassificationSchemaTrait
 {
-    // Adding the getBaseNodePath method.
-    use LomClassificationMetadataTrait;
-
     /**
-     * Get the classification source node's extract path.
+     * Get the default general source path
      *
      * @return array
      */
-    public function getNodePath()
-    {
-        return $this->getNodeRelativePath();
-    }
-
-    /**
-     * Get the classification source node's relative path.
-     *
-     * @return array
-     */
-    public function getNodeRelativePath()
+    public function getBaseNodePath()
     {
         return array(
-            $this->genericPathDefinition->getTaxonPath(),
-            $this->genericPathDefinition->getEntryPath(),
-            $this->genericPathDefinition->getStringPath(),
+            $this->genericPathDefinition->getLomPath(),
+            $this->genericPathDefinition->getClassificationPath(),
+            $this->genericPathDefinition->getTaxonPathPath(),
         );
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getTaoPath()
-    {
-        return null;
     }
 }
