+++ model/schema/imsglobal/general/LomGeneralLanguageSchema.php
@@ -21,12 +21,12 @@
 namespace oat\taoLom\model\schema\imsglobal\general;
 
 
-use oat\taoLom\model\schema\LomMetadataAbstract;
+use oat\taoLom\model\schema\LomSchemaAbstract;
 
-class LomGeneralIdentifierMetadata extends LomMetadataAbstract
+class LomGeneralLanguageSchema extends LomSchemaAbstract
 {
     // Adding the getBaseNodePath method.
-    use LomGeneralMetadataTrait;
+    use LomGeneralSchemaTrait;
 
     /**
      * Get the classification source node's extract path.
@@ -46,19 +46,18 @@ class LomGeneralIdentifierMetadata extends LomMetadataAbstract
     public function getNodeRelativePath()
     {
         return [
-            $this->genericPathDefinition->getIdentifierPath(),
-            $this->genericPathDefinition->getEntryPath(),
+            $this->genericPathDefinition->getLanguagePath(),
             $this->genericPathDefinition->getStringPath(),
         ];
     }
 
     /**
-     * Returns the general identifier place in the TAO system.
+     * Returns the general language place in the TAO system.
      *
      * @return string
      */
     public function getTaoPath()
     {
-        return $this->taoPathDefinition->getGeneralIdentifier();
+        return $this->taoPathDefinition->getGeneralLanguage();
     }
 }
